Refactor: split utils.field.vector into several modules.
