+++ nodes/field/rotation_field.py
@@ -6,7 +6,8 @@ from bpy.props import FloatProperty, EnumProperty, BoolProperty
 from sverchok.node_tree import SverchCustomTreeNode
 from sverchok.data_structure import updateNode, zip_long_repeat
 
-from sverchok.utils.field.vector import (SvAverageVectorField, SvRotationVectorField, SvSelectVectorField)
+from sverchok.utils.field.vector import (SvRotationVectorField)
+from sverchok.utils.field.vector_operations import (SvAverageVectorField, SvSelectVectorField)
 from sverchok.utils.math import all_falloff_types, falloff_array
 
 class SvRotationFieldNode(SverchCustomTreeNode, bpy.types.Node):
