+++ utils/field/scalar.py
@@ -6,15 +6,10 @@
 # License-Filename: LICENSE
 
 import numpy as np
-from math import copysign, sqrt, sin, cos, atan2, acos, pi
+from math import copysign, sqrt, sin, cos, atan2, acos
 
-from mathutils import Matrix, Vector
-from mathutils import kdtree
-from mathutils import bvhtree
-
-from sverchok.utils.math import from_cylindrical, from_spherical, to_cylindrical, to_spherical, np_dot
-from sverchok.utils.geom import LineEquation, CircleEquation3D
-from sverchok.utils.kdtree import SvKdTree
+from sverchok.utils.math import to_cylindrical, to_spherical
+from sverchok.utils.field.voronoi import SvVoronoiFieldData
 
 ##################
 #                #
@@ -135,52 +130,6 @@ class SvScalarFieldLambda(SvScalarField):
             V = self.in_field.evaluate(x, y, z)
         return self.function(x, y, z, V)
 
-class SvScalarFieldPointDistance(SvScalarField):
-    def __init__(self, center, metric='EUCLIDEAN', falloff=None, power=2):
-        self.center = center
-        self.falloff = falloff
-        self.metric = metric
-        self.power = power
-        self.__description__ = "Distance from {}".format(tuple(center))
-
-    def evaluate_grid(self, xs, ys, zs):
-        x0, y0, z0 = tuple(self.center)
-        xs = xs - x0
-        ys = ys - y0
-        zs = zs - z0
-        points = np.stack((xs, ys, zs))
-        if self.metric == 'EUCLIDEAN':
-            norms = np.linalg.norm(points, axis=0)
-        elif self.metric == 'CHEBYSHEV':
-            norms = np.max(np.abs(points), axis=0)
-        elif self.metric == 'MANHATTAN':
-            norms = np.sum(np.abs(points), axis=0)
-        elif self.metric == 'CUSTOM':
-            norms = np.linalg.norm(points, axis=0, ord=self.power)
-        else:
-            raise Exception('Unknown metric')
-        if self.falloff is not None:
-            result = self.falloff(norms)
-            return result
-        else:
-            return norms
-
-    def evaluate(self, x, y, z):
-        point = np.array([x, y, z]) - self.center
-        if self.metric == 'EUCLIDEAN':
-            norm = np.linalg.norm(point)
-        elif self.metric == 'CHEBYSHEV':
-            norm = np.max(np.abs(point))
-        elif self.metric == 'MANHATTAN':
-            norm = np.sum(np.abs(point))
-        elif self.metric == 'CUSTOM':
-            norm = np.linalg.norm(point, ord=self.power)
-        else:
-            raise Exception('Unknown metric')
-        if self.falloff is not None:
-            return self.falloff(np.array([norm]))[0]
-        else:
-            return norm
 
 class SvScalarFieldBinOp(SvScalarField):
     def __init__(self, field1, field2, function):
@@ -351,263 +300,6 @@ class SvMergedScalarField(SvScalarField):
             raise Exception("unsupported operation")
         return value
 
-class SvKdtScalarField(SvScalarField):
-    __description__ = "KDT"
-
-    def __init__(self, vertices=None, kdt=None, falloff=None, power=2):
-        self.falloff = falloff
-        if kdt is not None:
-            self.kdt = kdt
-        elif vertices is not None:
-            self.kdt = SvKdTree.new(SvKdTree.best_available_implementation(), vertices, power=power)
-        else:
-            raise Exception("Either kdt or vertices must be provided")
-
-    def evaluate(self, x, y, z):
-        nearest, i, distance = self.kdt.query(np.array([x,y,z]))
-        if self.falloff is not None:
-            value = self.falloff(np.array([distance]))[0]
-            return value
-        else:
-            return distance
-
-    def evaluate_grid(self, xs, ys, zs):
-        points = np.stack((xs, ys, zs)).T
-        locs, idxs, distances = self.kdt.query_array(points)
-        if self.falloff is not None:
-            result = self.falloff(distances)
-            return result
-        else:
-            return distances
-
-class SvLineAttractorScalarField(SvScalarField):
-    __description__ = "Line Attractor"
-
-    def __init__(self, center, direction, falloff=None):
-        self.center = center
-        self.direction = direction
-        self.falloff = falloff
-
-    def evaluate(self, x, y, z):
-        vertex = np.array([x,y,z])
-        direction = self.direction
-        to_center = self.center - vertex
-        projection = np.dot(to_center, direction) * direction / np.dot(direction, direction)
-        dv = to_center - projection
-        return np.linalg.norm(dv)
-
-    def evaluate_grid(self, xs, ys, zs):
-        direction = self.direction
-        direction2 = np.dot(direction, direction)
-        points = np.stack((xs, ys, zs)).T
-        to_center = self.center - points
-        dot = (to_center * direction).sum(axis=1)
-        projections = (dot * direction[np.newaxis].T / direction2).T
-        vectors = to_center - projections
-        norms = np.linalg.norm(vectors, axis=1)
-
-        if self.falloff is not None:
-            result = self.falloff(norms)
-            return result
-        else:
-            return norms
-
-class SvPlaneAttractorScalarField(SvScalarField):
-    __description__ = "Plane Attractor"
-
-    def __init__(self, center, direction, falloff=None):
-        self.center = center
-        self.direction = direction
-        self.falloff = falloff
-
-    def evaluate(self, x, y, z):
-        vertex = np.array([x,y,z])
-        direction = self.direction
-        to_center = self.center - vertex
-        projection = np.dot(to_center, direction) * direction / np.dot(direction, direction)
-        return np.linalg.norm(projection)
-
-    def evaluate_grid(self, xs, ys, zs):
-        direction = self.direction
-        direction2 = np.dot(direction, direction)
-
-        def func(vertex):
-            to_center = self.center - vertex
-            projection = np.dot(to_center, direction) * direction / direction2
-            return np.linalg.norm(projection)
-
-        points = np.stack((xs, ys, zs)).T
-        norms = np.vectorize(func, signature='(3)->()')(points)
-        if self.falloff is not None:
-            result = self.falloff(norms)
-            return result
-        else:
-            return norms
-
-class SvCircleAttractorScalarField(SvScalarField):
-    __description__ = "Circle Attractor"
-
-    def __init__(self, center, radius, normal, falloff=None):
-        self.circle = CircleEquation3D.from_center_radius_normal(center, radius, normal)
-        self.falloff = falloff
-
-    def evaluate(self, x, y, z):
-        v = np.array([x,y,z])
-        projection = self.circle.get_projections([v])[0]
-        distance = np.linalg.norm(v - projection)
-        if self.fallof is not None:
-            return self.falloff(np.array([distance]))[0]
-        else:
-            return distance
-
-    def evaluate_grid(self, xs, ys, zs):
-        vs = np.stack((xs, ys, zs)).T
-        projections = self.circle.get_projections(vs)
-        distances = np.linalg.norm(vs - projections, axis=1)
-        if self.falloff is not None:
-            return self.falloff(distances)
-        else:
-            return distances
-
-class SvBvhAttractorScalarField(SvScalarField):
-    __description__ = "BVH Attractor (faces)"
-
-    def __init__(self, bvh=None, verts=None, faces=None, falloff=None, signed=False):
-        self.falloff = falloff
-        self.signed = signed
-        if bvh is not None:
-            self.bvh = bvh
-        elif verts is not None and faces is not None:
-            self.bvh = bvhtree.BVHTree.FromPolygons(verts, faces)
-        else:
-            raise Exception("Either bvh or verts and faces must be provided!")
-
-    def evaluate(self, x, y, z):
-        nearest, normal, idx, distance = self.bvh.find_nearest((x,y,z))
-        if self.signed:
-            sign = (Vector((x,y,z)) - nearest).dot(normal)
-            sign = copysign(1, sign)
-        else:
-            sign = 1
-        value = sign * distance
-        if self.falloff is None:
-            return value
-        else:
-            return self.falloff(np.array([value]))[0]
-
-    def evaluate_grid(self, xs, ys, zs):
-        def find(v):
-            nearest, normal, idx, distance = self.bvh.find_nearest(v)
-            if nearest is None:
-                raise Exception("No nearest point on mesh found for vertex %s" % v)
-            if self.signed:
-                sign = (v - nearest).dot(normal)
-                sign = copysign(1, sign)
-            else:
-                sign = 1
-            return sign * distance
-
-        points = np.stack((xs, ys, zs)).T
-        norms = np.vectorize(find, signature='(3)->()')(points)
-        if self.falloff is not None:
-            result = self.falloff(norms)
-            return result
-        else:
-            return norms
-
-class SvBvhEdgesAttractorScalarField(SvScalarField):
-    __description__ = "BVH Attractor (edges)"
-
-    def __init__(self, verts, edges, falloff=None):
-        self.verts = verts
-        self.edges = edges
-        self.falloff = falloff
-        self.bvh = self._make_bvh(verts, edges)
-
-    def _make_bvh(self, verts, edges):
-        faces = [(i1, i2, i1) for i1, i2 in edges]
-        return bvhtree.BVHTree.FromPolygons(verts, faces)
-
-    def evaluate(self, x, y, z):
-        nearest, normal, idx, distance = self.bvh.find_nearest((x,y,z))
-        if self.falloff is None:
-            return distance
-        else:
-            return self.falloff(np.array([distance]))[0]
-
-    def evaluate_grid(self, xs, ys, zs):
-        def find(v):
-            nearest, normal, idx, distance = self.bvh.find_nearest(v)
-            return distance
-
-        points = np.stack((xs, ys, zs)).T
-        norms = np.vectorize(find, signature='(3)->()')(points)
-        if self.falloff is not None:
-            result = self.falloff(norms)
-            return result
-        else:
-            return norms
-
-class SvEdgeAttractorScalarField(SvScalarField):
-    __description__ = "Edge attractor"
-
-    def __init__(self, v1, v2, falloff=None):
-        self.falloff = falloff
-        self.v1 = Vector(v1)
-        self.v2 = Vector(v2)
-    
-    def evaluate(self, x, y, z):
-        v = Vector([x,y,z])
-        dv1 = (v - self.v1).length
-        dv2 = (v - self.v2).length
-        if dv1 > dv2:
-            distance_to_nearest = dv2
-            nearest_vert = self.v2
-            another_vert = self.v1
-        else:
-            distance_to_nearest = dv1
-            nearest_vert = self.v1
-            another_vert = self.v2
-        edge = another_vert - nearest_vert
-        to_nearest = v - nearest_vert
-        if to_nearest.length == 0:
-            return 0
-        angle = edge.angle(to_nearest)
-        if angle > pi/2:
-            distance = distance_to_nearest
-        else:
-            distance = LineEquation.from_two_points(self.v1, self.v2).distance_to_point(v)
-        if self.falloff is not None:
-            value = self.falloff(np.array([distance]))[0]
-            return value
-        else:
-            return distance
-    
-    def evaluate_grid(self, xs, ys, zs):
-        n = len(xs)
-        vs = np.stack((xs, ys, zs)).T
-        v1 = np.array(self.v1)
-        v2 = np.array(self.v2)    
-        dv1 = vs - v1
-        dv2 = vs - v2
-        edge = v2 - v1
-        dot1 = (dv1 * edge).sum(axis=1)
-        dot2 = -(dv2 * edge).sum(axis=1)
-        v1_is_nearest = (dot1 < 0)
-        v2_is_nearest = (dot2 < 0)
-        at_edge = np.logical_not(np.logical_or(v1_is_nearest, v2_is_nearest))
-
-        distances = np.empty((n,))
-        distances[v1_is_nearest] = np.linalg.norm(dv1[v1_is_nearest], axis=1)
-        distances[v2_is_nearest] = np.linalg.norm(dv2[v2_is_nearest], axis=1)
-        distances[at_edge] = LineEquation.from_two_points(self.v1, self.v2).distance_to_points(vs[at_edge])
-
-        if self.falloff is not None:
-            distances = self.falloff(distances)
-            return distances
-        else:
-            return distances
-
 class SvVectorScalarFieldComposition(SvScalarField):
     __description__ = "Composition"
 
@@ -623,76 +315,6 @@ class SvVectorScalarFieldComposition(SvScalarField):
     def evaluate_grid(self, xs, ys, zs):
         vx1, vy1, vz1 = self.vfield.evaluate_grid(xs, ys, zs)
         return self.sfield.evaluate_grid(vx1, vy1, vz1)
-
-class SvVectorFieldDivergence(SvScalarField):
-    def __init__(self, field, step):
-        self.field = field
-        self.step = step
-        self.__description__ = "Div({})".format(field)
-
-    def evaluate(self, x, y, z):
-        step = self.step
-        xs_dx_plus, _, _ = self.field.evaluate(x+step,y,z)
-        xs_dx_minus, _, _ = self.field.evaluate(x-step,y,z)
-        _, ys_dy_plus, _ = self.field.evaluate(x, y+step, z)
-        _, ys_dy_minus, _ = self.field.evaluate(x, y-step, z)
-        _, _, zs_dz_plus = self.field.evaluate(x, y, z+step)
-        _, _, zs_dz_minus = self.field.evaluate(x, y, z-step)
-
-        dx_dx = (xs_dx_plus - xs_dx_minus) / (2*step)
-        dy_dy = (ys_dy_plus - ys_dy_minus) / (2*step)
-        dz_dz = (zs_dz_plus - zs_dz_minus) / (2*step)
-
-        return dx_dx + dy_dy + dz_dz
-    
-    def evaluate_grid(self, xs, ys, zs):
-        step = self.step
-        xs_dx_plus, _, _ = self.field.evaluate_grid(xs+step, ys,zs)
-        xs_dx_minus, _, _ = self.field.evaluate_grid(xs-step,ys,zs)
-        _, ys_dy_plus, _ = self.field.evaluate_grid(xs, ys+step, zs)
-        _, ys_dy_minus, _ = self.field.evaluate_grid(xs, ys-step, zs)
-        _, _, zs_dz_plus = self.field.evaluate_grid(xs, ys, zs+step)
-        _, _, zs_dz_minus = self.field.evaluate_grid(xs, ys, zs-step)
-
-        dx_dx = (xs_dx_plus - xs_dx_minus) / (2*step)
-        dy_dy = (ys_dy_plus - ys_dy_minus) / (2*step)
-        dz_dz = (zs_dz_plus - zs_dz_minus) / (2*step)
-
-        return dx_dx + dy_dy + dz_dz
-
-class SvScalarFieldLaplacian(SvScalarField):
-    def __init__(self, field, step):
-        self.field = field
-        self.step = step
-        self.__description__ = "Laplace({})".format(field)
-
-    def evaluate(self, x, y, z):
-        step = self.step
-        v_dx_plus = self.field.evaluate(x+step,y,z)
-        v_dx_minus = self.field.evaluate(x-step,y,z)
-        v_dy_plus = self.field.evaluate(x, y+step, z)
-        v_dy_minus = self.field.evaluate(x, y-step, z)
-        v_dz_plus = self.field.evaluate(x, y, z+step)
-        v_dz_minus = self.field.evaluate(x, y, z-step)
-        v0 = self.field.evaluate(x, y, z)
-
-        sides = v_dx_plus + v_dx_minus + v_dy_plus + v_dy_minus + v_dz_plus + v_dz_minus
-        result = (sides - 6*v0) / (8 * step * step * step)
-        return result
-    
-    def evaluate_grid(self, xs, ys, zs):
-        step = self.step
-        v_dx_plus = self.field.evaluate_grid(xs+step, ys,zs)
-        v_dx_minus = self.field.evaluate_grid(xs-step,ys,zs)
-        v_dy_plus = self.field.evaluate_grid(xs, ys+step, zs)
-        v_dy_minus = self.field.evaluate_grid(xs, ys-step, zs)
-        v_dz_plus = self.field.evaluate_grid(xs, ys, zs+step)
-        v_dz_minus = self.field.evaluate_grid(xs, ys, zs-step)
-        v0 = self.field.evaluate_grid(xs, ys, zs)
-
-        sides = v_dx_plus + v_dx_minus + v_dy_plus + v_dy_minus + v_dz_plus + v_dz_minus
-        result = (sides - 6*v0) / (8 * step * step * step)
-        return result
 
 class ScalarFieldCurvatureCalculator(object):
     # Ref.: Curvature formulas for implicit curves and surfaces // Ron Goldman // doi:10.1016/j.cagd.2005.06.005
