+++ utils/field/vector.py
@@ -54,6 +54,32 @@ class SvVectorField(object):
     def to_absolute(self):
         return SvAbsoluteVectorField(self)
 
+class SvAbsoluteVectorField(SvVectorField):
+    def __init__(self, field):
+        self.field = field
+        self.__description__ = "Absolute({})".format(field)
+
+    def evaluate(self, x, y, z):
+        r = self.field.evaluate(x, y, z)
+        return r + np.array([x, y, z])
+
+    def evaluate_grid(self, xs, ys, zs):
+        rxs, rys, rzs = self.field.evaluate_grid(xs, ys, zs)
+        return rxs + xs, rys + ys, rzs + zs
+
+class SvRelativeVectorField(SvVectorField):
+    def __init__(self, field):
+        self.field = field
+        self.__description__ = "Relative({})".format(field)
+
+    def evaluate(self, x, y, z):
+        r = self.field.evaluate(x, y, z)
+        return r - np.array([x, y, z])
+
+    def evaluate_grid(self, xs, ys, zs):
+        rxs, rys, rzs = self.field.evaluate_grid(xs, ys, zs)
+        return rxs - xs, rys - ys, rzs - zs
+
 class SvMatrixVectorField(SvVectorField):
 
     def __init__(self, matrix):
@@ -88,66 +114,6 @@ class SvConstantVectorField(SvVectorField):
         rz = np.full_like(zs, z)
         return rx, ry, rz
 
-class SvComposedVectorField(SvVectorField):
-    def __init__(self, coords, sfield1, sfield2, sfield3):
-        self.coords = coords
-        self.sfield1 = sfield1
-        self.sfield2 = sfield2
-        self.sfield3 = sfield3
-        self.__description__ = "{}({}, {}, {})".format(coords, sfield1, sfield2, sfield3)
-
-    def evaluate(self, x, y, z):
-        v1 = self.sfield1.evaluate(x, y, z)
-        v2 = self.sfield2.evaluate(x, y, z)
-        v3 = self.sfield3.evaluate(x, y, z)
-        if self.coords == 'XYZ':
-            return np.array((v1, v2, v3))
-        elif self.coords == 'CYL':
-            return np.array(from_cylindrical(v1, v2, v3, mode='radians'))
-        else: # SPH:
-            return np.array(from_spherical(v1, v2, v3, mode='radians'))
-
-    def evaluate_grid(self, xs, ys, zs):
-        v1s = self.sfield1.evaluate_grid(xs, ys, zs)
-        v2s = self.sfield2.evaluate_grid(xs, ys, zs)
-        v3s = self.sfield3.evaluate_grid(xs, ys, zs)
-        if self.coords == 'XYZ':
-            return v1s, v2s, v3s
-        elif self.coords == 'CYL':
-            vectors = np.stack((v1s, v2s, v3s)).T
-            vectors = np.apply_along_axis(lambda v: np.array(from_cylindrical(*tuple(v), mode='radians')), 1, vectors).T
-            return vectors[0], vectors[1], vectors[2]
-        else: # SPH:
-            vectors = np.stack((v1s, v2s, v3s)).T
-            vectors = np.apply_along_axis(lambda v: np.array(from_spherical(*tuple(v), mode='radians')), 1, vectors).T
-            return vectors[0], vectors[1], vectors[2]
-
-class SvAbsoluteVectorField(SvVectorField):
-    def __init__(self, field):
-        self.field = field
-        self.__description__ = "Absolute({})".format(field)
-
-    def evaluate(self, x, y, z):
-        r = self.field.evaluate(x, y, z)
-        return r + np.array([x, y, z])
-
-    def evaluate_grid(self, xs, ys, zs):
-        rxs, rys, rzs = self.field.evaluate_grid(xs, ys, zs)
-        return rxs + xs, rys + ys, rzs + zs
-
-class SvRelativeVectorField(SvVectorField):
-    def __init__(self, field):
-        self.field = field
-        self.__description__ = "Relative({})".format(field)
-
-    def evaluate(self, x, y, z):
-        r = self.field.evaluate(x, y, z)
-        return r - np.array([x, y, z])
-
-    def evaluate_grid(self, xs, ys, zs):
-        rxs, rys, rzs = self.field.evaluate_grid(xs, ys, zs)
-        return rxs - xs, rys - ys, rzs - zs
-
 class SvVectorFieldLambda(SvVectorField):
 
     __description__ = "Formula"
@@ -178,108 +144,6 @@ class SvVectorFieldLambda(SvVectorField):
             V = self.in_field.evaluate(x, y, z)
         return np.array(self.function(x, y, z, V))
 
-class SvVectorFieldBinOp(SvVectorField):
-    def __init__(self, field1, field2, function):
-        self.function = function
-        self.field1 = field1
-        self.field2 = field2
-        self.__description__ = f"<BinOp ({field1}, {field2})>"
-
-    def evaluate(self, x, y, z):
-        return self.function(self.field1.evaluate(x, y, z), self.field2.evaluate(x, y, z))
-
-    def evaluate_grid(self, xs, ys, zs):
-        def func(xs, ys, zs):
-            vx1, vy1, vz1 = self.field1.evaluate_grid(xs, ys, zs)
-            vx2, vy2, vz2 = self.field2.evaluate_grid(xs, ys, zs)
-            R = self.function(np.array([vx1, vy1, vz1]), np.array([vx2, vy2, vz2]))
-            return R[0], R[1], R[2]
-        return np.vectorize(func, signature="(m),(m),(m)->(m),(m),(m)")(xs, ys, zs)
-
-class SvAverageVectorField(SvVectorField):
-
-    def __init__(self, fields):
-        self.fields = fields
-        self.__description__ = "Average"
-
-    def evaluate(self, x, y, z):
-        vectors = np.array([field.evaluate(x, y, z) for field in self.fields])
-        return np.mean(vectors, axis=0)
-
-    def evaluate_grid(self, xs, ys, zs):
-        def func(xs, ys, zs):
-            data = []
-            for field in self.fields:
-                vx, vy, vz = field.evaluate_grid(xs, ys, zs)
-                vectors = np.stack((vx, vy, vz)).T
-                data.append(vectors)
-            data = np.array(data)
-            mean = np.mean(data, axis=0).T
-            return mean[0], mean[1], mean[2]
-        return np.vectorize(func, signature="(m),(m),(m)->(m),(m),(m)")(xs, ys, zs)
-
-class SvVectorFieldCrossProduct(SvVectorField):
-    def __init__(self, field1, field2):
-        self.field1 = field1
-        self.field2 = field2
-        self.__description__ = "{} x {}".format(field1, field2)
-
-    def evaluate(self, x, y, z):
-        v1 = self.field1.evaluate(x, y, z)
-        v2 = self.field2.evaluate(x, y, z)
-        return np.cross(v1, v2)
-
-    def evaluate_grid(self, xs, ys, zs):
-        vx1, vy1, vz1 = self.field1.evaluate_grid(xs, ys, zs)
-        vx2, vy2, vz2 = self.field2.evaluate_grid(xs, ys, zs)
-        vectors1 = np.stack((vx1, vy1, vz1)).T
-        vectors2 = np.stack((vx2, vy2, vz2)).T
-        R = np.cross(vectors1, vectors2).T
-        return R[0], R[1], R[2]
-
-class SvVectorFieldMultipliedByScalar(SvVectorField):
-    def __init__(self, vector_field, scalar_field):
-        self.vector_field = vector_field
-        self.scalar_field = scalar_field
-        self.__description__ = "{} * {}".format(scalar_field, vector_field)
-
-    def evaluate(self, x, y, z):
-        scalar = self.scalar_field.evaluate(x, y, z)
-        vector = self.vector_field.evaluate(x, y, z)
-        return scalar * vector
-
-    def evaluate_grid(self, xs, ys, zs):
-        def product(xs, ys, zs):
-            scalars = self.scalar_field.evaluate_grid(xs, ys, zs)
-            vx, vy, vz = self.vector_field.evaluate_grid(xs, ys, zs)
-            vectors = np.stack((vx, vy, vz))
-            R = (scalars * vectors)
-            return R[0], R[1], R[2]
-        return np.vectorize(product, signature="(m),(m),(m)->(m),(m),(m)")(xs, ys, zs)
-
-class SvVectorFieldsLerp(SvVectorField):
-
-    def __init__(self, vfield1, vfield2, scalar_field):
-        self.vfield1 = vfield1
-        self.vfield2 = vfield2
-        self.scalar_field = scalar_field
-        self.__description__ = "Lerp"
-
-    def evaluate(self, x, y, z):
-        scalar = self.scalar_field.evaluate(x, y, z)
-        vector1 = self.vfield1.evaluate(x, y, z)
-        vector2 = self.vfield2.evaluate(x, y, z)
-        return (1 - scalar) * vector1 + scalar * vector2
-
-    def evaluate_grid(self, xs, ys, zs):
-            scalars = self.scalar_field.evaluate_grid(xs, ys, zs)
-            vx1, vy1, vz1 = self.vfield1.evaluate_grid(xs, ys, zs)
-            vectors1 = np.stack((vx1, vy1, vz1))
-            vx2, vy2, vz2 = self.vfield2.evaluate_grid(xs, ys, zs)
-            vectors2 = np.stack((vx2, vy2, vz2))
-            R = (1 - scalars) * vectors1 + scalars * vectors2
-            return R[0], R[1], R[2]
-
 class SvNoiseVectorField(SvVectorField):
     def __init__(self, noise_type, seed):
         self.noise_type = noise_type
@@ -298,358 +162,6 @@ class SvNoiseVectorField(SvVectorField):
             return r[0], r[1], r[2]
         vectors = np.stack((xs,ys,zs)).T
         return np.vectorize(mk_noise, signature="(3)->(),(),()")(vectors)
-
-class SvKdtVectorField(SvVectorField):
-
-    def __init__(self, vertices=None, kdt=None, falloff=None, negate=False, power=2):
-        self.falloff = falloff
-        self.negate = negate
-        if kdt is not None:
-            self.kdt = kdt
-        elif vertices is not None:
-            self.kdt = SvKdTree.new(SvKdTree.best_available_implementation(), vertices, power=power)
-        else:
-            raise Exception("Either kdt or vertices must be provided")
-        self.__description__ = "KDT Attractor"
-
-    def evaluate(self, x, y, z):
-        nearest, i, distance = self.kdt.query(np.array([x, y, z]))
-        vector = nearest - np.array([x, y, z])
-        if self.falloff is not None:
-            value = self.falloff(np.array([distance]))[0]
-            if self.negate:
-                value = - value
-            norm = np.linalg.norm(vector)
-            return value * vector / norm
-        else:
-            if self.negate:
-                return - vector
-            else:
-                return vector
-
-    def evaluate_grid(self, xs, ys, zs):
-        points = np.stack((xs, ys, zs)).T
-        locs, idxs, distances = self.kdt.query_array(points)
-        vectors = locs - points
-        if self.negate:
-            vectors = - vectors
-        if self.falloff is not None:
-            norms = np.linalg.norm(vectors, axis=1, keepdims=True)
-            lens = self.falloff(norms)
-            nonzero = (norms > 0)[:,0]
-            lens = self.falloff(norms)
-            vectors[nonzero] = vectors[nonzero] / norms[nonzero]
-            R = (lens * vectors).T
-            return R[0], R[1], R[2]
-        else:
-            R = vectors.T
-            return R[0], R[1], R[2]
-
-class SvVectorFieldPointDistance(SvVectorField):
-    def __init__(self, center, metric='EUCLIDEAN', falloff=None, power=2):
-        self.center = center
-        self.falloff = falloff
-        self.metric = metric
-        self.power = power
-        self.__description__ = "Distance from {}".format(tuple(center))
-
-    def evaluate_grid(self, xs, ys, zs):
-        x0, y0, z0 = tuple(self.center)
-        xs = x0 - xs
-        ys = y0 - ys
-        zs = z0 - zs
-        vectors = np.stack((xs, ys, zs))
-        if self.metric == 'EUCLIDEAN':
-            norms = np.linalg.norm(vectors, axis=0)
-        elif self.metric == 'CHEBYSHEV':
-            norms = np.max(np.abs(vectors), axis=0)
-        elif self.metric == 'MANHATTAN':
-            norms = np.sum(np.abs(vectors), axis=0)
-        elif self.metric == 'CUSTOM':
-            norms = np.linalg.norm(vectors, axis=0, ord=self.power)
-        else:
-            raise Exception('Unknown metric')
-        if self.falloff is not None:
-            lens = self.falloff(norms)
-            R = lens * vectors / norms
-        else:
-            R = vectors
-        return R[0], R[1], R[2]
-
-    def evaluate(self, x, y, z):
-        point = np.array([x, y, z]) - self.center
-        if self.metric == 'EUCLIDEAN':
-            norm = np.linalg.norm(point)
-        elif self.metric == 'CHEBYSHEV':
-            norm = np.max(point)
-        elif self.metric == 'MANHATTAN':
-            norm = np.sum(np.abs(point))
-        elif self.metric == 'CUSTOM':
-            norm = np.linalg.norm(point, ord=self.power)
-        else:
-            raise Exception('Unknown metric')
-        if self.falloff is not None:
-            value = self.falloff(np.array([norm]))[0]
-            return value * point / norm
-        else:
-            return point
-
-class SvLineAttractorVectorField(SvVectorField):
-
-    def __init__(self, center, direction, falloff=None):
-        self.center = center
-        self.direction = direction
-        self.falloff = falloff
-        self.__description__ = "Line Attractor"
-
-    def evaluate(self, x, y, z):
-        vertex = np.array([x,y,z])
-        direction = self.direction
-        to_center = self.center - vertex
-        projection = np.dot(to_center, direction) * direction / np.dot(direction, direction)
-        dv = to_center - projection
-        if self.falloff is not None:
-            norm = np.linalg.norm(dv)
-            dv = self.falloff(norm) * dv / norm
-        return dv
-
-    def evaluate_grid(self, xs, ys, zs):
-        direction = self.direction
-        direction2 = np.dot(direction, direction)
-
-        def func(vertex):
-            to_center = self.center - vertex
-            projection = np.dot(to_center, direction) * direction / direction2
-            dv = to_center - projection
-            return dv
-
-        points = np.stack((xs, ys, zs)).T
-        vectors = np.vectorize(func, signature='(3)->(3)')(points)
-        if self.falloff is not None:
-            norms = np.linalg.norm(vectors, axis=1, keepdims=True)
-            nonzero = (norms > 0)[:,0]
-            lens = self.falloff(norms)
-            vectors[nonzero] = vectors[nonzero] / norms[nonzero][:,0][np.newaxis].T
-            R = (lens * vectors).T
-            return R[0], R[1], R[2]
-        else:
-            R = vectors.T
-            return R[0], R[1], R[2]
-
-class SvPlaneAttractorVectorField(SvVectorField):
-
-    def __init__(self, center, direction, falloff=None):
-        self.center = center
-        self.direction = direction
-        self.falloff = falloff
-        self.__description__ = "Plane Attractor"
-
-    def evaluate(self, x, y, z):
-        vertex = np.array([x,y,z])
-        direction = self.direction
-        to_center = self.center - vertex
-        dv = np.dot(to_center, direction) * direction / np.dot(direction, direction)
-        if self.falloff is not None:
-            norm = np.linalg.norm(dv)
-            dv = self.falloff(norm) * dv / norm
-        return dv
-
-    def evaluate_grid(self, xs, ys, zs):
-        direction = self.direction
-        direction2 = np.dot(direction, direction)
-
-        def func(vertex):
-            to_center = self.center - vertex
-            projection = np.dot(to_center, direction) * direction / direction2
-            return projection
-
-        points = np.stack((xs, ys, zs)).T
-        vectors = np.vectorize(func, signature='(3)->(3)')(points)
-        if self.falloff is not None:
-            norms = np.linalg.norm(vectors, axis=1, keepdims=True)
-            lens = self.falloff(norms)
-            nonzero = (norms > 0)[:,0]
-            vectors[nonzero] = vectors[nonzero] / norms[nonzero][:,0][np.newaxis].T
-            R = (lens * vectors).T
-            return R[0], R[1], R[2]
-        else:
-            R = vectors.T
-            return R[0], R[1], R[2]
-
-class SvCircleAttractorVectorField(SvVectorField):
-    __description__ = "Circle Attractor"
-
-    def __init__(self, center, radius, normal, falloff=None):
-        self.circle = CircleEquation3D.from_center_radius_normal(center, radius, normal)
-        self.falloff = falloff
-
-    def evaluate(self, x, y, z):
-        v = np.array([x,y,z])
-        projection = self.circle.get_projections([v])[0]
-        vector = projection - v
-        if self.fallof is not None:
-            new_len = self.falloff(np.array([distance]))[0]
-            norm = np.linalg.norm(vector)
-            return new_len * vector / norm
-        else:
-            return vector
-
-    def evaluate_grid(self, xs, ys, zs):
-        vs = np.stack((xs, ys, zs)).T
-        projections = self.circle.get_projections(vs)
-        vectors = projections - vs
-        if self.falloff is not None:
-            norms = np.linalg.norm(vectors, axis=1, keepdims=True)
-            lens = self.falloff(norms)
-            nonzero = (norms > 0)[:,0]
-            vectors[nonzero] = vectors[nonzero] / norms[nonzero][:,0][np.newaxis].T
-            R = (lens * vectors).T
-            return R[0], R[1], R[2]
-        else:
-            R = vectors.T
-            return R[0], R[1], R[2]
-
-class SvEdgeAttractorVectorField(SvVectorField):
-    __description__ = "Edge attractor"
-
-    def __init__(self, v1, v2, falloff=None):
-        self.falloff = falloff
-        self.v1 = Vector(v1)
-        self.v2 = Vector(v2)
-
-    def evaluate(self, x, y, z):
-        v = Vector([x,y,z])
-        dv1 = (v - self.v1).length
-        dv2 = (v - self.v2).length
-        if dv1 > dv2:
-            distance_to_nearest = dv2
-            nearest_vert = self.v2
-            another_vert = self.v1
-        else:
-            distance_to_nearest = dv1
-            nearest_vert = self.v1
-            another_vert = self.v2
-        edge = another_vert - nearest_vert
-        to_nearest = v - nearest_vert
-        if to_nearest.length == 0:
-            return 0
-        angle = edge.angle(to_nearest)
-        if angle > pi/2:
-            distance = distance_to_nearest
-            vector = - to_nearest
-        else:
-            vector = LineEquation.from_two_points(self.v1, self.v2).projection_of_points(v)
-            distance = vector.length
-            vector = np.array(vector)
-        if self.falloff is not None:
-            return self.falloff(distance) * vector / distance
-        else:
-            return vector
-
-    def evaluate_grid(self, xs, ys, zs):
-        n = len(xs)
-        vs = np.stack((xs, ys, zs)).T
-        v1 = np.array(self.v1)
-        v2 = np.array(self.v2)
-        dv1s = np.linalg.norm(vs - v1, axis=1)
-        dv2s = np.linalg.norm(vs - v2, axis=1)
-        v1_is_nearest = (dv1s < dv2s)
-        v2_is_nearest = np.logical_not(v1_is_nearest)
-        nearest_verts = np.empty_like(vs)
-        other_verts = np.empty_like(vs)
-        nearest_verts[v1_is_nearest] = v1
-        nearest_verts[v2_is_nearest] = v2
-        other_verts[v1_is_nearest] = v2
-        other_verts[v2_is_nearest] = v1
-
-        to_nearest = vs - nearest_verts
-
-        edges = other_verts - nearest_verts
-        dot = (to_nearest * edges).sum(axis=1)
-        at_edge = (dot > 0)
-        at_vertex = np.logical_not(at_edge)
-        at_v1 = np.logical_and(at_vertex, v1_is_nearest)
-        at_v2 = np.logical_and(at_vertex, v2_is_nearest)
-
-        line = LineEquation.from_two_points(self.v1, self.v2)
-
-        vectors = np.empty((n,3))
-        vectors[at_edge] = line.projection_of_points(vs[at_edge]) - vs[at_edge]
-        vectors[at_v1] = v1 - vs[at_v1]
-        vectors[at_v2] = v2 - vs[at_v2]
-
-        if self.falloff is not None:
-            norms = np.linalg.norm(vectors, axis=1, keepdims=True)
-            lens = self.falloff(norms)
-            nonzero = (norms > 0)[:,0]
-            vectors[nonzero] = vectors[nonzero] / norms[nonzero][:,0][np.newaxis].T
-            R = (lens * vectors).T
-            return R[0], R[1], R[2]
-        else:
-            R = vectors.T
-            return R[0], R[1], R[2]
-
-class SvBvhAttractorVectorField(SvVectorField):
-
-    def __init__(self, bvh=None, verts=None, faces=None, falloff=None, use_normal=False, signed_normal=False):
-        self.falloff = falloff
-        self.use_normal = use_normal
-        self.signed_normal = signed_normal
-        if bvh is not None:
-            self.bvh = bvh
-        elif verts is not None and faces is not None:
-            self.bvh = bvhtree.BVHTree.FromPolygons(verts, faces)
-        else:
-            raise Exception("Either bvh or verts and faces must be provided!")
-        self.__description__ = "BVH Attractor"
-
-    def evaluate(self, x, y, z):
-        vertex = Vector((x,y,z))
-        nearest, normal, idx, distance = self.bvh.find_nearest(vertex)
-        if self.use_normal:
-            if self.signed_normal:
-                sign = (v - nearest).dot(normal)
-                sign = copysign(1, sign)
-            else:
-                sign = 1
-            return sign * np.array(normal)
-        else:
-            dv = np.array(nearest - vertex)
-            if self.falloff is not None:
-                norm = np.linalg.norm(dv)
-                len = self.falloff(norm)
-                dv = len * dv
-                return dv
-            else:
-                return dv
-
-    def evaluate_grid(self, xs, ys, zs):
-        def find(v):
-            nearest, normal, idx, distance = self.bvh.find_nearest(v)
-            if nearest is None:
-                raise Exception("No nearest point on mesh found for vertex %s" % v)
-            if self.use_normal:
-                if self.signed_normal:
-                    sign = (v - nearest).dot(normal)
-                    sign = copysign(1, sign)
-                else:
-                    sign = 1
-                return sign * np.array(normal)
-            else:
-                return np.array(nearest) - v
-
-        points = np.stack((xs, ys, zs)).T
-        vectors = np.vectorize(find, signature='(3)->(3)')(points)
-        if self.falloff is not None:
-            norms = np.linalg.norm(vectors, axis=1, keepdims=True)
-            nonzero = (norms > 0)[:,0]
-            lens = self.falloff(norms)
-            vectors[nonzero] = vectors[nonzero] / norms[nonzero]
-            R = (lens * vectors).T
-            return R[0], R[1], R[2]
-        else:
-            R = vectors.T
-            return R[0], R[1], R[2]
 
 class SvRotationVectorField(SvVectorField):
 
@@ -767,254 +279,6 @@ class SvTaperVectorField(SvVectorField):
     def evaluate(self, x, y, z):
         rxs, rys, rzs = self.evaluate_grid([x], [y], [z])
         return rxs[0], rys[0], rzs[0]
-
-class SvSelectVectorField(SvVectorField):
-    def __init__(self, fields, mode):
-        self.fields = fields
-        self.mode = mode
-        self.__description__ = "{}({})".format(mode, fields)
-
-    def evaluate(self, x, y, z):
-        vectors = [field.evaluate(x, y, z) for field in self.fields]
-        vectors = np.array(vectors)
-        norms = np.linalg.norm(vectors, axis=1)
-        if self.mode == 'MIN':
-            selected = np.argmin(norms)
-        else: # MAX
-            selected = np.argmax(norms)
-        return vectors[selected]
-
-    def evaluate_grid(self, xs, ys, zs):
-        n = len(xs)
-        vectors = [field.evaluate_grid(xs, ys, zs) for field in self.fields]
-        vectors = np.stack(vectors)
-        vectors = np.transpose(vectors, axes=(2,0,1))
-        norms = np.linalg.norm(vectors, axis=2)
-        if self.mode == 'MIN':
-            selected = np.argmin(norms, axis=1)
-        else: # MAX
-            selected = np.argmax(norms, axis=1)
-        all_points = list(range(n))
-        vectors = vectors[all_points, selected, :]
-        return vectors.T
-
-class SvVectorFieldCartesianFilter(SvVectorField):
-    def __init__(self, field, use_x=True, use_y=True, use_z=True):
-        self.field = field
-        self.use_x = use_x
-        self.use_y = use_y
-        self.use_z = use_z
-        desc = ""
-        if use_x:
-            desc = desc + 'X'
-        if use_y:
-            desc = desc + 'Y'
-        if use_z:
-            desc = desc + 'Z'
-        self.__description__ = f"Filter[{desc}]({field})"
-
-    def evaluate_grid(self, xs, ys, zs):
-        vxs, vys, vzs = self.field.evaluate_grid(xs, ys, zs)
-        if not self.use_x:
-            vxs[:] = 0
-        if not self.use_y:
-            vys[:] = 0
-        if not self.use_z:
-            vzs[:] = 0
-        return vxs, vys, vzs
-
-class SvVectorFieldCylindricalFilter(SvVectorField):
-    def __init__(self, field, use_rho=True, use_phi=True, use_z=True):
-        self.field = field
-        self.use_rho = use_rho
-        self.use_phi = use_phi
-        self.use_z = use_z
-
-    def evaluate_grid(self, xs, ys, zs):
-        pts = np.stack((xs,ys,zs)).T
-        vxs, vys, vzs = self.field.evaluate_grid(xs, ys, zs)
-        vectors = np.stack((vxs,vys,vzs)).T
-        s_rho, s_phi, s_z = to_cylindrical_np(pts.T, mode='radians')
-        v_rho, v_phi, v_z = to_cylindrical_np((pts + vectors).T, mode='radians')
-        if not self.use_rho:
-            v_rho = s_rho
-        if not self.use_phi:
-            v_phi = s_phi
-        if not self.use_z:
-            v_z = s_z
-        v_x, v_y, v_z = from_cylindrical_np(v_rho, v_phi, v_z, mode='radians')
-        return (v_x - xs), (v_y - ys), (v_z - zs)
-
-class SvVectorFieldSphericalFilter(SvVectorField):
-    def __init__(self, field, use_rho=True, use_phi=True, use_theta=True):
-        self.field = field
-        self.use_rho = use_rho
-        self.use_phi = use_phi
-        self.use_theta = use_theta
-
-    def evaluate_grid(self, xs, ys, zs):
-        pts = np.stack((xs,ys,zs)).T
-        vxs, vys, vzs = self.field.evaluate_grid(xs, ys, zs)
-        vectors = np.stack((vxs,vys,vzs)).T
-        s_rho, s_phi, s_theta = to_spherical_np(pts.T, mode='radians')
-        v_rho, v_phi, v_theta = to_spherical_np((pts + vectors).T, mode='radians')
-        if not self.use_rho:
-            v_rho = s_rho
-        if not self.use_phi:
-            v_phi = s_phi
-        if not self.use_theta:
-            v_theta = s_theta
-        v_x, v_y, v_z = from_spherical_np(v_rho, v_phi, v_theta, mode='radians')
-        return (v_x - xs), (v_y - ys), (v_z - zs)
-
-class SvVectorFieldTangent(SvVectorField):
-
-    def __init__(self, field1, field2):
-        self.field1 = field1
-        self.field2 = field2
-        self.__description__ = "Tangent"
-
-    def evaluate(self, x, y, z):
-        v1 = self.field1.evaluate(x,y,z)
-        v2 = self.field2.evaluate(x,y,z)
-        projection = np.dot(v1, v2) * v2 / np.dot(v2, v2)
-        return projection
-
-    def evaluate_grid(self, xs, ys, zs):
-        vx1, vy1, vz1 = self.field1.evaluate_grid(xs, ys, zs)
-        vx2, vy2, vz2 = self.field2.evaluate_grid(xs, ys, zs)
-        vectors1 = np.stack((vx1, vy1, vz1)).T
-        vectors2 = np.stack((vx2, vy2, vz2)).T
-
-        def project(v1, v2):
-            projection = np.dot(v1, v2) * v2 / np.dot(v2, v2)
-            vx, vy, vz = projection
-            return vx, vy, vz
-
-        return np.vectorize(project, signature="(3),(3)->(),(),()")(vectors1, vectors2)
-
-class SvVectorFieldCotangent(SvVectorField):
-
-    def __init__(self, field1, field2):
-        self.field1 = field1
-        self.field2 = field2
-        self.__description__ = "Cotangent"
-
-    def evaluate(self, x, y, z):
-        v1 = self.field1.evaluate(x,y,z)
-        v2 = self.field2.evaluate(x,y,z)
-        projection = np.dot(v1, v2) * v2 / np.dot(v2, v2)
-        return v1 - projection
-
-    def evaluate_grid(self, xs, ys, zs):
-        vx1, vy1, vz1 = self.field1.evaluate_grid(xs, ys, zs)
-        vx2, vy2, vz2 = self.field2.evaluate_grid(xs, ys, zs)
-        vectors1 = np.stack((vx1, vy1, vz1)).T
-        vectors2 = np.stack((vx2, vy2, vz2)).T
-
-        def project(v1, v2):
-            projection = np.dot(v1, v2) * v2 / np.dot(v2, v2)
-            coprojection = v1 - projection
-            vx, vy, vz = coprojection
-            return vx, vy, vz
-
-        return np.vectorize(project, signature="(3),(3)->(),(),()")(vectors1, vectors2)
-
-class SvVectorFieldComposition(SvVectorField):
-
-    def __init__(self, field1, field2):
-        self.field1 = field1
-        self.field2 = field2
-        self.__description__ = "Composition"
-
-    def evaluate(self, x, y, z):
-        x1, y1, z1 = self.field1.evaluate(x,y,z)
-        v2 = self.field2.evaluate(x1,y1,z1)
-        return v2
-
-    def evaluate_grid(self, xs, ys, zs):
-        r = self.field1.evaluate_grid(xs, ys, zs)
-        vx1, vy1, vz1 = r
-        return self.field2.evaluate_grid(vx1, vy1, vz1)
-
-class SvPreserveCoordinateField(SvVectorField):
-    def __init__(self, field, axis):
-        self.field = field
-        self.axis = axis
-
-    def evaluate(self, x, y, z):
-        xyz = self.field.evaluate(x, y, z)
-        xyz = np.array(xyz)
-        xyz[self.axis] = [x, y, z][self.axis]
-        return xyz
-
-    def evaluate_grid(self, xs, ys, zs):
-        r = self.field.evaluate_grid(xs, ys, zs)
-        r = np.array(r)
-        r[self.axis] = [xs, ys, zs][self.axis]
-        return r
-
-class SvScalarFieldGradient(SvVectorField):
-    def __init__(self, field, step):
-        self.field = field
-        self.step = step
-        self.__description__ = "Grad({})".format(field)
-
-    def evaluate(self, x, y, z):
-        return self.field.gradient([x, y, z], step=self.step)
-
-    def evaluate_grid(self, xs, ys, zs):
-        return self.field.gradient_grid(xs, ys, zs, step=self.step)
-
-class SvVectorFieldRotor(SvVectorField):
-    def __init__(self, field, step):
-        self.field = field
-        self.step = step
-        self.__description__ = "Rot({})".format(field)
-
-    def evaluate(self, x, y, z):
-        step = self.step
-        _, y_dx_plus, z_dx_plus = self.field.evaluate(x+step,y,z)
-        _, y_dx_minus, z_dx_minus = self.field.evaluate(x-step,y,z)
-        x_dy_plus, _, z_dy_plus = self.field.evaluate(x, y+step, z)
-        x_dy_minus, _, z_dy_minus = self.field.evaluate(x, y-step, z)
-        x_dz_plus, y_dz_plus, _ = self.field.evaluate(x, y, z+step)
-        x_dz_minus, y_dz_minus, _ = self.field.evaluate(x, y, z-step)
-
-        dy_dx = (y_dx_plus - y_dx_minus) / (2*step)
-        dz_dx = (z_dx_plus - z_dx_minus) / (2*step)
-        dx_dy = (x_dy_plus - x_dy_minus) / (2*step)
-        dz_dy = (z_dy_plus - z_dy_minus) / (2*step)
-        dx_dz = (x_dz_plus - x_dz_minus) / (2*step)
-        dy_dz = (y_dz_plus - y_dz_minus) / (2*step)
-
-        rx = dz_dy - dy_dz
-        ry = - (dz_dx - dx_dz)
-        rz = dy_dx - dx_dy
-
-        return np.array([rx, ry, rz])
-
-    def evaluate_grid(self, xs, ys, zs):
-        step = self.step
-        _, y_dx_plus, z_dx_plus = self.field.evaluate_grid(xs+step,ys,zs)
-        _, y_dx_minus, z_dx_minus = self.field.evaluate_grid(xs-step,ys,zs)
-        x_dy_plus, _, z_dy_plus = self.field.evaluate_grid(xs, ys+step, zs)
-        x_dy_minus, _, z_dy_minus = self.field.evaluate_grid(xs, ys-step, zs)
-        x_dz_plus, y_dz_plus, _ = self.field.evaluate_grid(xs, ys, zs+step)
-        x_dz_minus, y_dz_minus, _ = self.field.evaluate_grid(xs, ys, zs-step)
-
-        dy_dx = (y_dx_plus - y_dx_minus) / (2*step)
-        dz_dx = (z_dx_plus - z_dx_minus) / (2*step)
-        dx_dy = (x_dy_plus - x_dy_minus) / (2*step)
-        dz_dy = (z_dy_plus - z_dy_minus) / (2*step)
-        dx_dz = (x_dz_plus - x_dz_minus) / (2*step)
-        dy_dz = (y_dz_plus - y_dz_minus) / (2*step)
-
-        rx = dz_dy - dy_dz
-        ry = - (dz_dx - dx_dz)
-        rz = dy_dx - dx_dy
-        R = np.stack((rx, ry, rz))
-        return R[0], R[1], R[2]
 
 class SvBendAlongCurveField(SvVectorField):
 
@@ -1280,3 +544,4 @@ class SvScalarFieldCurveMap(SvVectorField):
         else: # NORMAL
             vectors = self.curve.main_normal_array(ts)
         return vectors[:,0], vectors[:,1], vectors[:,2]
+
